Enable including libraries with a preamble via the HTML script tag
Mutating the self variable with Object.create on the window global
makes objects included on js.context within Dart unavailable in the
browser environment. Fix this issue by only preventing encapsulation
via Object.create when necessary and not in the browser.

Additionally, browsers tend to not have CommonJS globals available. To
prevent ReferenceErrors when running in the browser, check if these
globals are available before setting them on self.